+++ pkg/failurehandler/apps.go
@@ -1,7 +1,6 @@
 package failurehandler
 
 import (
-	"context"
 	"fmt"
 
 	"github.com/giantswarm/clustertest"
@@ -15,8 +14,11 @@ import (
 
 // AppIssues produces debugging information from app-operator (on the MC) and chart-operator (on the WC)
 // This function will log out the status of the deployments, any related Events found and the last 25 lines of logs from the pods.
-func AppIssues(ctx context.Context, framework *clustertest.Framework, cluster *application.Cluster) FailureHandler {
+func AppIssues(framework *clustertest.Framework, cluster *application.Cluster) FailureHandler {
 	return Wrap(func() {
+		ctx, cancel := newContext()
+		defer cancel()
+
 		logger.Log("Attempting to get debug info for App related failure")
 
 		// Note: This doesn't check the MCs own app-operator so issues with bundles won't be surfaced by this.